+++ drivers/stepper/adi_tmc/tmc51xx/CMakeLists.txt
@@ -1,4 +1,4 @@
 # SPDX-FileCopyrightText: Copyright (c) 2025 Dipak Shetty
 # SPDX-License-Identifier: Apache-2.0
 
-zephyr_library_sources_ifdef(CONFIG_STEPPER_ADI_TMC51XX tmc51xx.c tmc51xx_spi.c tmc51xx_uart.c)
+zephyr_library_sources_ifdef(CONFIG_STEPPER_ADI_TMC51XX tmc51xx.c)
